wip(#1104): extract lower/expr_call/url_search_params
